Delete DummyJSONComment type
