--- src/lib/types/DummyJSONComment.ts
@@ -1,9 +1,0 @@
-type DummyJSONComment = {
-  id: number;
-  body: string;
-  user: { id: number };
-  postId: number;
-  likes: number;
-};
-
-export default DummyJSONComment;
