+++ src/lib/types/index.ts
@@ -12,6 +12,5 @@ export { type default as TPost } from './TPost';
 export { type default as TPosts } from './TPosts';
 export { type default as TReaction } from './TReaction';
 
-export { type default as DummyJSONComment } from './DummyJSONComment';
 export { type default as DummyJSONPost } from './DummyJSONPost';
 export { type default as DummyJSONUser } from './DummyJSONUser';
